LayoutTests/media/video-duration-known-after-eos.html

<!DOCTYPE HTML>

<html>
    <head>
        <title>local video</title>

        <script src=media-file.js></script>
        <script src=video-test.js></script>

        <script>
            function ended()
            {
                testExpected("video.duration", 0, '>');
                endTest();
            }

            function start()
            {
                findMediaElement();

                video.addEventListener("ended", ended);

                video.src = findMediaFile("video", "content/test");
                consoleWrite("");
                run("video.play()");
            }

        </script>
    </head>
    <body>
        <video controls autobuffer></video>
        <p>Tests that duration is known after playback ended.</p>
        <script>start()</script>
    </body>
</html>
