LayoutTests/media/video-duration-known-after-eos-expected.txt

Tests that duration is known after playback ended.


RUN(video.play())
EXPECTED (video.duration > '0') OK
END OF TEST
